requirements.txt
# Big Data & Traitement des Données
pyspark
pandas
numpy
scipy
pyarrow
requests
markdown
beautifulsoup4
pymongo
azure-storage-blob
langdetect
transformers
torch
matplotlib
scikit-learn
mlflow
tensorflow
streamlit